requirements.txt
absl-py==2.1.0
accelerate==1.6.0
addict==2.4.0
argcomplete==3.3.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
ccimport==0.4.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cumm-cu124==0.7.11
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.33.1
docopt==0.6.2
easydict==1.13
easydict==1.13
einops==0.8.1
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.13.1
fire==0.7.0
Flask==3.0.3
fonttools==4.56.0
fsspec==2024.6.1
grpcio==1.70.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio==2.37.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
lazy_loader==0.4
line_profiler==4.2.0
llvmlite==0.44.0
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
narwhals==1.30.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
ninja==1.11.1.3
numba==0.61.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open3d==0.19.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pccm==0.4.16
pcdet==0.6.0+8caccce
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.0.1
pipx==1.6.0
platformdirs==4.2.2
plotly==6.0.0
portalocker==3.1.1
prompt_toolkit==3.0.50
protobuf==6.30.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
Pygments==2.19.1
PyMCubes==0.1.6
pyparsing==3.2.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rpds-py==0.23.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
SharedArray==3.2.4
SharedArray==3.2.4
six==1.17.0
soupsieve==2.7
spconv-cu124==2.3.8
stack_data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2025.2.18
tifffile==2025.2.18
timm==1.0.14
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.5.1+cu124
torch_cluster==1.6.3+pt25cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trimesh==4.6.2
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
userpath==1.9.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
yarg==0.1.9
zipp==3.21.0
